+++ nacc_attribute_deriver/attributes/nacc/modules/uds/form_a1.py
@@ -141,6 +141,17 @@ class UDSFormA1Attribute(UDSAttributeCollection):
 
         return source == 4 or sourcenw == 2
 
+    def _create_naccpaff(self) -> int:
+        """Creates NACCPAFF - Previously affiliated subject.
+
+        Always set as long as being an affiliate is ever true.
+        """
+        naccpaff = self.__subject_derived.get_cross_sectional_value("naccpaff", int)
+        if naccpaff == 1:
+            return 1
+
+        return 1 if self._create_affiliate() else 0
+
     def _create_educ(self) -> Optional[int]:
         """UDS education level."""
         return self.uds.get_value("educ", int)
@@ -148,6 +159,7 @@ class UDSFormA1Attribute(UDSAttributeCollection):
     def _create_prespart(self) -> Optional[int]:
         """Presumed participation.
 
-        Used for NACCACTV.
+        Used for NACCACTV. Usually only provided at first visit so need
+        to carry over as a working variable
         """
         return self.uds.get_value("prespart", int)